+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 8.6
 -----
+- Improve cache invalidation for Playlist Count and Artwork [#3927](https://github.com/Automattic/pocket-casts-ios/pull/3927)
 
 - [Dynamic Type] Update UpNext [#3918](https://github.com/Automattic/pocket-casts-ios/pull/3918)
 - [TestFlight only] Add a new VoiceBoost solution designed to maintain more consistent audio levels. [#3915](https://github.com/Automattic/pocket-casts-ios/pull/3915)